requirements.txt
Django==3.2.9
Pillow==8.4.0
django-crispy-forms==1.13.0